Adding useAppManager hook, adding doc-blocks in app component, extracted interfaces to files
