+++ src/app/interfaces/user/user.interface.ts
@@ -1,0 +1,10 @@
+interface IUser {
+  firstName: string;
+  LastName: string;
+}
+export interface IUserData {
+  user: IUser;
+  courses: {
+    totalProgressBar: number;
+  };
+}
